// packages/rive/src/types.ts
export type RefCallback<T> = (instance: T | null) => void;

export type CanvasElementProps = {
	className?: string;
	style?: Record<string, string | number | undefined> | null;
	children?: OctaneNode;
} & Record<string, unknown>;

export type UseRiveParameters =
	| (Partial<Omit<RiveParameters, 'canvas'>> & {
			onRiveReady?: (rive: Rive) => void;
	  })
	| null;

export type UseRiveOptions = {
	useDevicePixelRatio: boolean;
	customDevicePixelRatio: number;
	fitCanvasToArtboardHeight: boolean;
	useOffscreenRenderer: boolean;
	shouldResizeCanvasToContainer: boolean;
	shouldUseIntersectionObserver?: boolean;
};
// ...
export type RiveState = {
	canvas: HTMLCanvasElement | null;
	container: HTMLElement | null;
	setCanvasRef: RefCallback<HTMLCanvasElement>;
	setContainerRef: RefCallback<HTMLElement>;
	rive: Rive | null;
	RiveComponent: (props: CanvasElementProps) => OctaneNode;
};